+++ include/file_readers/obj_parser.hpp
@@ -27,7 +27,7 @@
    
    Returns true if the operation was successful
 */
-bool parse_obj_file(const char* file_name, const std::optional<size_t> default_texture_index,
+bool parse_obj_file(const char* file_name, const std::optional<unsigned int> default_texture_index,
    std::vector<const object*>& obj_set,
    std::vector<wrapper<material>>& material_wrapper_set,
    std::vector<wrapper<texture>>& texture_wrapper_set,
