+++ include/scene/objects/box.hpp
@@ -28,7 +28,7 @@ class box : public object {
         
         /* Main constructor */
         box(const rt::vector& center, const rt::vector& n1, const rt::vector& n2,
-            const real l1, const real l2, const real l3, const size_t material_index);
+            const real l1, const real l2, const real l3, const unsigned int material_index);
 
         /* Constructor used for bounding boxes: no material and no index */
         box(const rt::vector& center, const rt::vector& n1, const rt::vector& n2,
