+++ src/file_readers/bmp_reader.cpp
@@ -151,9 +151,9 @@ bool read_bmp(const char* file_name, std::vector<std::vector<rt::color>>& data) 
         }
 
         unsigned int index = 0;
-        for (size_t j = 0; j < bmpheight; j++) {
-            const size_t indexj = bmpheight - j - 1;
-            for (size_t i = 0; i < bmpwidth; i++) {
+        for (unsigned int j = 0; j < bmpheight; j++) {
+            const unsigned int indexj = bmpheight - j - 1;
+            for (unsigned int i = 0; i < bmpwidth; i++) {
                 const real b = buffer[index];
                 index++;
                 const real g = buffer[index];
@@ -243,7 +243,7 @@ bool write_bmp(const char* file_name, std::vector<std::vector<rt::color>>& data,
     const unsigned int width = data.size();
     const unsigned int height = data[0].size();
     bool padding = false;
-    size_t p = (3 * width) % 4;
+    unsigned int p = (3 * width) % 4;
     if (p != 0) {
         padding = true;
         p = 4 - p;
@@ -360,7 +360,7 @@ bool write_bmp(const char* file_name, std::vector<std::vector<rt::color>>& data,
 
         /* Padding bytes at the end of each line (if padding = true) */
         char padding_zeroes[MAX_PADDING];
-        for (size_t k = 0; k < p; k++) {
+        for (unsigned int k = 0; k < p; k++) {
             padding_zeroes[k] = 0;
         }
 
@@ -369,9 +369,9 @@ bool write_bmp(const char* file_name, std::vector<std::vector<rt::color>>& data,
         const double inv255 = 1.0 / 255.0;
         // const double inv255invN = invN * inv255;
 
-        for (size_t j = 0; j < height; j++) {
-            const size_t indexj = height - j - 1;
-            for (size_t i = 0; i < width; i++) {
+        for (unsigned int j = 0; j < height; j++) {
+            const unsigned int indexj = height - j - 1;
+            for (unsigned int i = 0; i < width; i++) {
                 const rt::color& c = data[i][indexj];
                 if (gamma_enabled) {
                     const unsigned char r = pow((std::min((c.get_red()   * invN), 255.0) * inv255), gamma) * 255.0;
