+++ src/scene/objects/triangle.cpp
@@ -42,7 +42,7 @@ std::pair<real, int> set_up_det(const rt::vector& v1, const rt::vector& v2) {
 
 // Constructor from three points without vertex normals
 triangle::triangle(const rt::vector& p0, const rt::vector& p1, const rt::vector& p2, 
-    const size_t material_index, const std::optional<texture_info>& info)
+    const unsigned int material_index, const std::optional<texture_info>& info)
 
     : object(p0, material_index, info) {
 
@@ -65,7 +65,7 @@ triangle::triangle(const rt::vector& p0, const rt::vector& p1, const rt::vector&
 // Constructor from three points with vertex normals
 triangle::triangle(const rt::vector& p0, const rt::vector& p1, const rt::vector& p2,
     const rt::vector& vn0init, const rt::vector& vn1, const rt::vector& vn2,
-    const size_t material_index, const std::optional<texture_info>& info)
+    const unsigned int material_index, const std::optional<texture_info>& info)
 
     : object(p0, material_index, info), vn0(vn0init.unit())
     //, vn1(vn1.unit()), vn2(vn2.unit())
@@ -87,7 +87,7 @@ triangle::triangle(const rt::vector& p0, const rt::vector& p1, const rt::vector&
 
 // Constructor from three points with normal mapping enabled
 triangle::triangle(const rt::vector& p0, const rt::vector& p1, const rt::vector& p2,
-    const size_t material_index, const std::optional<texture_info>& info, const bool normal_mapping)
+    const unsigned int material_index, const std::optional<texture_info>& info, const bool normal_mapping)
 
     : object(p0, material_index, info) {
     
@@ -142,7 +142,7 @@ triangle::triangle(const rt::vector& p0, const rt::vector& p1, const rt::vector&
 // Constructor from three points with vertex normals and normal mapping enabled
 triangle::triangle(const rt::vector& p0, const rt::vector& p1, const rt::vector& p2,
     const rt::vector& vn0init, const rt::vector& vn1, const rt::vector& vn2,
-    const size_t material_index, const std::optional<texture_info>& info, const bool normal_mapping)
+    const unsigned int material_index, const std::optional<texture_info>& info, const bool normal_mapping)
 
     : object(p0, material_index, info), vn0(vn0init.unit())
     //, vn1(vn1.unit()), vn2(vn2.unit())
